+++ sakura_core/_os/Clipboard.h
@@ -35,7 +35,7 @@ public:
 	// インターフェース
 	void Empty(); // クリップボードを空にする
 	void Close(); // クリップボードを閉じる
-	bool SetText(const wchar_t* pData, int nDataLen, bool bColumnSelect, bool bLineSelect, UINT uFormat = (UINT)-1);   // テキストを設定する
+	bool SetText(const wchar_t* pData, size_t nDataLen, bool bColumnSelect, bool bLineSelect, UINT uFormat = (UINT)-1);   // テキストを設定する
 	bool SetHtmlText(const NativeW& memBUf);
 	bool GetText(NativeW* pMemBuf, bool* pbColumnSelect, bool* pbLineSelect, const Eol& eol, UINT uGetFormat = (UINT)-1); // テキストを取得する
 	bool IsIncludeClipboradFormat(const wchar_t* pFormatName);
